// store/actions/types/cities.ts
export enum GetCitiesTypes {
  GET_CITIES_LOADING = "GET_CITIES_LOADING",
  GET_CITIES_SUCCESS = "GET_CITIES_SUCCESS",
  GET_CITIES_ERROR = "GET_CITIES_ERROR",
}

export enum CityOperationsTypes {
  OPERATION_CITY_LOADING = "OPERATION_CITY_LOADING",
  OPERATION_CITY_SUCCESS = "OPERATION_CITY_SUCCESS",
  OPERATION_CITY_ERROR = "OPERATION_CITY_ERROR",
}